Ensure it's an integer

+++ src/Util/Pagination.php
@@ -126,7 +126,7 @@ class Pagination
 
     public function getCurrentPage(): int
     {
-        return Input::get($this->getUrlParameter()) ?: 1;
+        return (int) (Input::get($this->getUrlParameter()) ?: 1);
     }
 
     /**
